[Dagger Refactor] Merge ModuleConstructoProxyGenerator into ModuleProxies.
This CL merges ModuleConstructoProxyGenerator into ModuleProxies to reduce the visibility of methods in ModuleProxies that are only used by the generator. This also consolidates all of the logic (both the generator and retriever) for the proxy into one class.

Note: Nothing except the top-level module references ModuleConstructorProxyGenerator directly.

RELNOTES=N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=272064871
